create functional chat based off group model
