+++ app/assets/javascripts/channels/rooms.coffee
@@ -6,8 +6,8 @@ $(document).on 'ready page:load', ->
     messages_to_bottom()
 
     App.global_chat = App.cable.subscriptions.create {
-        channel: "ChatRoomsChannel"
-        chat_room_id: messages.data('chat-room-id')
+        channel: "GroupsChannel"
+        group_id: messages.data('group-id')
       },
       connected: ->
         # Called when the subscription is ready for use on the server
@@ -19,15 +19,15 @@ $(document).on 'ready page:load', ->
         messages.append data['message']
         messages_to_bottom()
 
-      send_message: (message, chat_room_id) ->
-        @perform 'send_message', message: message, chat_room_id: chat_room_id
+      send_message: (message, group_id) ->
+        @perform 'send_message', message: message, group_id: group_id
 
 
     $('#new_message').submit (e) ->
       $this = $(this)
       textarea = $this.find('#message_body')
       if $.trim(textarea.val()).length > 1
-        App.global_chat.send_message textarea.val(), messages.data('chat-room-id')
+        App.global_chat.send_message textarea.val(), messages.data('group-id')
         textarea.val('')
       e.preventDefault()
       return false
