+++ app/controllers/messages_controller.rb
@@ -2,7 +2,6 @@ class MessagesController < ApplicationController
 
   def new
     @message = Message.new
-
   end
 
   def index
