--- app/helpers/chat_rooms_helper.rb
@@ -1,2 +1,0 @@
-module ChatRoomsHelper
-end
