+++ app/jobs/message_broadcast_job.rb
@@ -2,7 +2,7 @@ class MessageBroadcastJob < ApplicationJob
   queue_as :default
 
   def perform(message)
-    ActionCable.server.broadcast "chat_rooms_#{message.chat_room.id}_channel",
+    ActionCable.server.broadcast "groups_#{message.group.id}_channel",
     message: render_message(message)
   end
