--- app/models/chatlinkage.rb
@@ -1,4 +1,0 @@
-class Chatlinkage < ApplicationRecord
-  belongs_to :chat_room
-  belongs_to :user
-end
