+++ app/models/group.rb
@@ -1,8 +1,8 @@
 class Group < ApplicationRecord
   belongs_to :event
-  has_one :chat_room
   has_many :members
   has_many :users, through: :members
+  has_many :messages, dependent: :destroy
