+++ app/models/message.rb
@@ -1,13 +1,10 @@
 class Message < ApplicationRecord
   belongs_to :user
-  belongs_to :chat_room
-  has_many :comments
+  belongs_to :group
   validates :body, presence: true, length: {minimum: 2, maximum: 1000}
-
   after_create_commit { MessageBroadcastJob.perform_later(self) }
 
-def timestamp
-  created_at.strftime('%H:%M:%S %d %B %Y')
-end
-
+  def timestamp
+    created_at.strftime('%H:%M:%S %d %B %Y')
+  end
 end
