+++ app/models/user.rb
@@ -10,13 +10,11 @@ class User < ApplicationRecord
   has_many :events, through: :goings
   has_many :members
   has_many :groups, through: :members
-  has_many :chatlinkages
-  has_many :chat_rooms, through: :chatlinkages
+  has_many :messages, dependent: :destroy
 
 
   has_many :authored_conversations, class_name: 'Conversation', foreign_key: 'author_id'
   has_many :received_conversations, class_name: 'Conversation', foreign_key: 'receiver_id'
 
   has_many :personal_messages, dependent: :destroy
-  has_many :messages, dependent: :destroy
 end
