--- app/views/chat_rooms/_chat_room.html.erb
@@ -1,1 +1,0 @@
-<li><%= link_to "Enter #{chat_room.title}", event_group_chat_room_path(@event, @group, chat_room) %></li>
