--- app/views/chat_rooms/index.html.erb
@@ -1,7 +1,0 @@
-<h1>Chat rooms</h1>
-
-<%= link_to 'New chat room', new_event_group_chat_room_path(@event, @group), class: 'btn btn-primary' %>
-
-<ul>
-  <%= render @chat_rooms %>
-</ul>
