--- app/views/chat_rooms/new.html.erb
@@ -1,10 +1,0 @@
-<h1>Add chat room</h1>
-
-<%= form_for ([@group, @chat_room]), url: event_group_chat_rooms_path(@event, @group), method: :post  do |f| %>
-  <div class="form-group">
-    <%= f.label :title %>
-    <%= f.text_field :title, autofocus: true, class: 'form-control' %>
-  </div>
-
-  <%= f.submit "Add!", class: 'btn btn-primary' %>
-<% end %>
