--- app/views/chat_rooms/show.html.erb
@@ -1,1 +1,0 @@
-<%= render 'single_room' %>
