+++ app/views/groups/show.html.erb
@@ -1,1 +1,20 @@
 <%= render 'single_group', g: @group %>
+
+<h1><%= @group.name %></h1>
+
+<div id="messages" data-group-id="<%= @group.id %>">
+  <%= render @group.messages %>
+</div>
+
+<hr>
+
+<%= form_for @message, url: '#' do |f| %>
+  <%= hidden_field_tag 'group_id', @group.id %>
+  <div class="form-group">
+    <%= f.label :body %>
+    <%= f.text_area :body, class: 'form-control' %>
+    <small class="text-muted">From 2 to 1000 characters</small>
+  </div>
+
+ <%= f.submit "Post", class: 'btn btn-primary btn-lg' %>
+<% end %>
