--- db/migrate/20170515203146_create_chat_rooms.rb
@@ -1,9 +1,0 @@
-class CreateChatRooms < ActiveRecord::Migration[5.0]
-  def change
-    create_table :chat_rooms do |t|
-      t.string :title
-
-      t.timestamps
-    end
-  end
-end
