+++ db/migrate/20170516191932_create_messages.rb
@@ -1,0 +1,11 @@
+class CreateMessages < ActiveRecord::Migration[5.0]
+  def change
+    create_table :messages do |t|
+      t.text :body
+      t.references :user, foreign_key: true
+      t.references :group, foreign_key: true
+
+      t.timestamps
+    end
+  end
+end
