--- test/controllers/chat_rooms_controller_test.rb
@@ -1,7 +1,0 @@
-require 'test_helper'
-
-class ChatRoomsControllerTest < ActionDispatch::IntegrationTest
-  # test "the truth" do
-  #   assert true
-  # end
-end
