--- test/fixtures/chat_rooms.yml
@@ -1,13 +1,0 @@
-# Read about fixtures at http://api.rubyonrails.org/classes/ActiveRecord/FixtureSet.html
-
-one:
-  title: MyString
-  user: one
-  group: one
-  event: one
-
-two:
-  title: MyString
-  user: two
-  group: two
-  event: two
