--- test/models/chat_room_test.rb
@@ -1,7 +1,0 @@
-require 'test_helper'
-
-class ChatRoomTest < ActiveSupport::TestCase
-  # test "the truth" do
-  #   assert true
-  # end
-end
